+++ shared/api.js
@@ -5,6 +5,19 @@ async function getFromApi(api, url) {
   return data;
 }
 
+const getAllPages = async (api, url) => {
+  let hasMore = true;
+  let results = [];
+  while (hasMore) {
+    const { data } = await api.get(url);
+    results.push(...data.items);
+    hasMore = data.hasMore;
+    url = data.nextPage;
+  }
+  return results;
+};
+
 module.exports = {
   getFromApi,
+  getAllPages,
 };
